Prototype: refactor the app for one file per endpoint
see https://fastapi.tiangolo.com/tutorial/bigger-applications/ for instructions
